backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "SoluTrain API"
authors = ["Dor", "Omer Priel <omer.priel2@gmail.com>", "Ori Sharaby", "Ori Sharaby", "Stav Sharon"]
license = "MIT"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
pytest = "^7.4.1"
httpx = "^0.25.0"
pytest-mock = "^3.11.1"
python-dotenv = "^1.0.0"
psycopg = {extras = ["binary", "pool"], version = "^3.1.16"}
python-multipart = "^0.0.6"
pycryptodome = "^3.20.0"
pandas = "^2.2.2"
openpyxl = "^3.1.2"
pandas-stubs = "^2.2.1.240316"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.7.0"
flake8 = "^6.1.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']
line-length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
      \.git
    | \.mypy_cache
  )/
)
'''

[tool.isort]
line_length=120
multi_line_output=3
include_trailing_comma="True"
